backend/requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0

# Database
pymongo==4.5.0
mongoengine==0.27.0

# AI/ML Libraries
tensorflow>=2.16.0
scikit-learn>=1.3.0
pandas>=2.0.3
numpy>=1.24.3

# Cloud Provider SDKs
boto3==1.28.25  # AWS
azure-mgmt-compute==30.1.0  # Azure
google-cloud-compute==1.14.0  # GCP

# Data Processing
scipy==1.11.2
joblib==1.3.2

# Monitoring & Metrics
psutil==5.9.5
prometheus-client==0.17.1

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.7.0
flake8==6.1.0
pylint==2.17.5

# Security
cryptography==41.0.3
bcrypt==4.0.1

# API Documentation
flasgger==0.9.7.1

# Scheduling
APScheduler==3.10.4
